v3

+++ routes/index.js
@@ -48,30 +48,37 @@ var returnRouter = function(io) {
             var conv = cmp.conversion.split('/');
             
             var filePath = 'public/data/'+cmp.company+'/'+dirName+'/'+conv[0]+'-'+conv[1]+'.json';
-            var rawdata = fs.readFileSync(filePath);
-            var jsonData = JSON.parse(rawdata);
-            jsonData = ('Data' in jsonData) ? jsonData.Data : jsonData;
 
-            const timestamp = Math.floor(new Date() / 1000);
-            link = "https://min-api.cryptocompare.com/data/histo"+type+"?fsym="+conv[0]+"&tsym="+conv[1]+"&limit=2&toTs="+timestamp+"&e="+cmp.company;
-            
-            var ip = req.connection.remoteAddress;
-            
-            request.get({url: link, localAdress: ip}, function(error, request, body){
-              if(error){
-                callback(error, null);
-              }
-              var rates = JSON.parse(body);
-              rates = rates.Data;
-              if(rates.length > 0){
-                jsonData.push(rates[rates.length-1]);
-                fs.writeFileSync(filePath, JSON.stringify(jsonData));
+            if(fs.existsSync(filePath)){
+              var rawdata = fs.readFileSync(filePath);
+              var jsonData = JSON.parse(rawdata);
+              jsonData = ('Data' in jsonData) ? jsonData.Data : jsonData;
+  
+              const timestamp = Math.floor(new Date() / 1000);
+              link = "https://min-api.cryptocompare.com/data/histo"+type+"?fsym="+conv[0]+"&tsym="+conv[1]+"&limit=2&toTs="+timestamp+"&e="+cmp.company;
+              
+              var ip = req.connection.remoteAddress;
+              
+              request.get({url: link, localAdress: ip}, function(error, request, body){
+                if(error){
+                  callback(error, null);
+                }
+                var rates = JSON.parse(body);
+                rates = rates.Data;
+                if(rates.length > 0){
+                  jsonData.push(rates[rates.length-1]);
+                  fs.writeFileSync(filePath, JSON.stringify(jsonData));
+  
+                  myArray.push(jsonData);
+                }
+  
+                done();
+              });
 
-                myArray.push(jsonData);
-              }
-
+            }
+            else{
               done();
-            });
+            }
           },function(err){
               callback(null);
           });
